pyproject.toml
[tool.poetry]
name = "_openai"
version = "0.1.0"
description = ""
authors = ["yaaa3ser <yasserissa1411@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
openai = "^0.27.0"
python-decouple = "^3.8"
django-extensions = "^3.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
